+++ src/vector/common.jl
@@ -14,6 +14,11 @@ abstract type ScalarToScalarBijector end
 
 The same as `identity`.
 
+This is the appropriate scalar-to-scalar bijector for univariate distributions which have
+support over the entire real line. However, note that this can also be used in other
+contexts apart from univariate distributions. For example, `TypedIdentity()` is the correct
+implementation of `to_vec` and `from_vec` for `MvNormal` distributions.
+
 The problem with using `identity` as a bijector is that ChangesOfVariables.jl defines
 `with_logabsdet_jacobian(identity, x) = (x, zero(eltype(x)))`, which can fail if `eltype(x)`
 is not a number type! Implementing this allows us to shortcircuit that definition and return