+++ src/vector/transformed.jl
@@ -1,0 +1,11 @@
+# Note that to_vec, from_vec, vec_length, and optic_vec are inherited from the generic
+# UnivariateDistribution, MultivariateDistribution, and MatrixDistribution methods so
+# we don't need to define them here.
+for T in (B.UnivariateTransformed, B.MultivariateTransformed, B.MatrixTransformed)
+    @eval begin
+        to_linked_vec(td::$T) = to_linked_vec(td.dist) ∘ inverse(td.transform)
+        from_linked_vec(td::$T) = td.transform ∘ from_linked_vec(td.dist)
+        linked_vec_length(td::$T) = linked_vec_length(td.dist)
+        linked_optic_vec(td::$T) = fill(nothing, linked_vec_length(td))
+    end
+end